size_t type updates.

+++ code/TargetAnimation.cpp
@@ -100,8 +100,8 @@ void KeyIterator::operator ++()
 	// to our current position on the time line
 	double d0,d1;
 	
-	d0 = objPos->at      ( std::min<unsigned int> ( nextObjPos, objPos->size()-1)             ).mTime;
-	d1 = targetObjPos->at( std::min<unsigned int> ( nextTargetObjPos, targetObjPos->size()-1) ).mTime;	
+	d0 = objPos->at      ( std::min<size_t> ( nextObjPos, objPos->size()-1)             ).mTime;
+	d1 = targetObjPos->at( std::min<size_t> ( nextTargetObjPos, targetObjPos->size()-1) ).mTime;	
 	
 	// Easiest case - all are identical. In this
 	// case we don't need to interpolate so we can
